Move deduplication up to the generic level
Now that we're using SdmGoals, and we don't get the whole graph at once and updates are
more random in ordering, it's more common to get double-triggered for the same goal.

and, it was a bug in deploy, now that the status updates are outside the deploy call,
the deduping needs to be outside of that. Stuff was getting updated to success just because
a duplicate came in and didn't need to do anything.