Show deploy errors (#271)
* start

* labels

* attempt to guess edges

* comment out guess

* start

* refactor of SetGoalsToPending

* beginning graph play

* this is showing me a graph locally

* bug fix: thought everything was an automation client

promises are always truthy

* Add autofix for node build script

* Automatic de-linting
[atomist:auto-delint]

* Make verify-endpoint use generic

this will not work until OnSuccess does auto-complete-goal

* Improved formatting of graph

* improve generic goal handling, 1/n: comments

* graph play: logging

* 2/n: creds not needed for preconditionStatus

* 3/n: if preconditions failed, don't update status, just warn

The FailDownstream handler will update its status soon enough
This removes another dependency on GH statuses

* 4/n: extract precondition checks into function

* 5/n: bug fix, be careful on identifying goal

* 6/n: another bug fix, allow goalSet to be passed in

* 6/n: do 'pending' status via SdmGoal

* 7/n: move precondition check into only OnSuccess

Now that we distinguish 'requested' from 'planned', requested means GO

* 8/n: make it a special case if we are not handling updates

* 9/n: move common functions to another file

* 10/n: extract function

* 11/n: share code for updating goal

* 12/n: Use goal updates from verify

now finally, this should work with verify on the general case

* show deploy errors 1/p: give the things a name

so I can see them in the logs
also give them retries

* 2/p: Remove duplication, functional unit for goals

* fixup! start

* fixup! start

* 3/p: add logging to find failure

* 4/p: establish failure cascade deploy->endpoint

* 5/p: pull checkout into a function

* 6/p: remove split of executeDeploy to deploy

This also makes it not update its own deploy status
and use runWithLog; it's both simpler and supplies the error handling we want

* try to revert some

* Log to file locally and NOT in cluster

* don't always call logs 'deploy'

* Skip updating GitHub status for in_process goals

It stomps on successes

* debug logs

* Do not save logs in ConsoleLogger

That is what InMemoryProgressLog is for

* Throw error after reporting it

it is the surrounding handler's job to set the goal to failed

* Move deduplication up to the generic level

Now that we're using SdmGoals, and we don't get the whole graph at once and updates are
more random in ordering, it's more common to get double-triggered for the same goal.

and, it was a bug in deploy, now that the status updates are outside the deploy call,
the deduping needs to be outside of that. Stuff was getting updated to success just because
a duplicate came in and didn't need to do anything.

* Make Java stuff work again by making Deploy wait for Artifact

This is going to break Node again, and we need a solution for that.

* comment

* lint

* Make k8 deployers use functionalUnitForGoal

* Local deployment is Spring support

it doesn't matter where test/prod deploy is

* update graphviz test to match current goals

* logInterpreter is no longer optional to runWithLogs

* Automatic de-linting
[atomist:auto-delint]

* Autofix: TypeScript header

[atomist]

* Improve maven interpretation; makes tests pass

* Fix: runWithLog passes on ExecuteGoalResult

this fixes approval

* Autofix: tslint

[atomist]

* The payoff -- we don't need a hardcoded list of all goals

* store log in consoleLog

so we don't need the inMemory one then

* skip functional units for goals with nothing to do

Fingerprinters had this before. The others didn't, but should

* Autofix: tslint

[atomist]