Use CloudFoundryBlueGreenDeployer
I tested this several different ways in the actual application by running the goals. I ran into a few problems with getting it to run, but the main one was a 422 on app start because our PCF account was out of memory.
I'd like to modify the retry for that http call to give more info about the error in the log, and also not actually retry in that case.