Dispose and hooks (#320)
* Make undeploy required

* For every deploy goal, an undeploy goal

* Pass repoRef instead of artifact into finDeployments

that's the only bit anything was using. For undeploy, I need to call that without having an artifact

* Make findDeployments required

needed for undeploy

* Implement Undeploy goals

* We don't really need GitHubRepoRef

and one function is shorter

* Set rules for project disposal

* Set disposal goals based on defined rules,

triggered by a command.

* Command to delete the repository

it can't be done as an event, needs user token

* Warn if no disposal goals matched

* bug fixes. And, undeploy Node as well

* After undeploying, offer to delete the repository

* Offer to delete repo after undeploys complete

* How about just wait for the prod undeploy

let's avoid race conditions

* lint

* Remember to implement tthe otther goal

* unique name

* allow pre and post goal hooks

this allows to execute arbitary pre and post hooks in form of shell scripts sitting inside `.atomist` or the repo. Scripts need to be named `<environment>-<name>_<pre|post>-hook.sh`.

if the hooks exits with > 0 the goal will marked failed.

* delint

* allow execution of any executable file

* move to .atomist/hooks directory

* fix missing env vars

* Use ProgressLoader and same logs

This is a big change and needs tested with the smoke tests

* Autofix: TypeScript header

[atomist]

* Fix code reaction

and linting, unfortunately

* Fix retry

* cleanup

* lint

* merge errors

* Autofix: tslint

[atomist]