If the log is available online, don't show it
This removes all the places we were saying to please show the log
Instead if the log has a URL it will skip it, and if it does not,
we will send it to Slack as a snippet.

That way if Rolar is down we'll get the log in slack by default.

the log interpreters can still explicitly include or not include it.