unset 'no-unnecessary-callback-wrapper' as it introduces errors
This one autofixes [].map(s => stringify(s)) to [].map(stringify) which is incorrect
We have hit errors with this several times now, Rod's suffered from it too.
This rule is NOT OK because if the wrapped callback has optional arguments, map will
happily pass in the index for the second one, which is NOT THE SAME behavior as undefined.