Optionally, register the location of the registration
This enables autofixes to add that info to commit messages
which enables developers to change what an autofix does,
because they can find the code.

This is part of a delivery system that provides visibility into what
it is doing, part of development automation that lets devs craft their
own world.