Kill timed out container goals
The first container in container goals will exit when the SDM goal
timeout is exceeded.  This does not kill the pod.  Add a check for
this case in the job cleanup logic.  Increase frequency of job
cleanup.  Centralize lookup of SDM goal timeout and job TTL.

Only record pod spec in KubernetesGoalScheduler.

Do not perform sync in isolated goals.

Towards #816 #815